Deleted the first while, and started optional challenge 3
While loop: also changed the comments a bit to explain the algorithm.
Optional challenge 3: added a data structure for keeping track of the parking lot constraints, updated the constructor to allow constraints, added check for validity of states w.r.t. constraints and added a couple of tests for invalid states.